requirements.txt
anyio==4.4.0
beautifulsoup4==4.12.3
browser-cookie3==0.19.1
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jeepney==0.8.0
lxml==5.3.0
lz4==4.3.3
msal==1.31.0
pushbullet.py==0.12.0
pycparser==2.22
pycryptodomex==3.20.0
PyJWT==2.9.0
python-magic==0.4.27
requests==2.32.3
setuptools==72.1.0
sniffio==1.3.1
soupsieve==2.6
urllib3==2.2.2
websocket-client==1.8.0
wheel==0.43.0
